refactor: 修复 setPageResult

+++ easybbs-common/src/main/java/utils/SetPageUtils.java
@@ -6,9 +6,9 @@ import response.PageResult;
 public class SetPageUtils {
 
     public static void setPageResult(PageInfo page, PageResult result) {
-        result.setPageTotal(page.getPages());
+        result.setPageTotal((long) page.getPages());
         result.setPageNo(page.getPageNum());
-        result.setPageSize(page.getPageSize());
+        result.setPageSize((long) page.getPageSize());
         result.setTotalCount(page.getTotal());
         result.setList(page.getList());
     }
